Don't run delete SQL query if the list is empty

+++ lib/indx/objectstore_async.py
@@ -564,7 +564,8 @@ class ObjectStoreAsync:
 
                 if 'deleted' in diff:
                     urilist = diff['deleted']
-                    queries['prelatest']['queries'].append(("DELETE FROM wb_latest_vers USING wb_triples, wb_strings AS subjects WHERE subjects.id_string = wb_triples.subject AND wb_latest_vers.triple = wb_triples.id_triple AND subjects.string = ANY(%s)", [urilist]))
+                    if len(urilist) > 0:
+                        queries['prelatest']['queries'].append(("DELETE FROM wb_latest_vers USING wb_triples, wb_strings AS subjects WHERE subjects.id_string = wb_triples.subject AND wb_latest_vers.triple = wb_triples.id_triple AND subjects.string = ANY(%s)", [urilist]))
 
                     for uri in urilist:
                         queries['diff']['values'].append("(%s, %s, wb_get_string_id(%s), wb_get_string_id(%s), NULL, NULL)")
